+++ internal/pkg/middleware/requestlog/middleware.go
@@ -1,15 +1,18 @@
-package httpmiddleware
+package requestlog
 
 import (
 	"net/http"
 	"time"
 
-	"github.com/abyssparanoia/rapid-go/internal/pkg/log"
-	"github.com/go-chi/chi/middleware"
+	"github.com/abyssparanoia/rapid-go/internal/pkg/error/httperror"
+
+	"github.com/blendle/zapdriver"
+	"github.com/google/uuid"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
 )
+
+const producerID = "rapid-go"
 
 // HTTPMiddleware ... http middleware
 type HTTPMiddleware struct {
@@ -20,13 +23,15 @@ type HTTPMiddleware struct {
 func (m *HTTPMiddleware) Handle(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		start := time.Now()
-		var requestID string
-		if reqID := r.Context().Value(middleware.RequestIDKey); reqID != nil {
-			requestID = reqID.(string)
-		}
+
+		operationID := uuid.New()
+
+		m.logger.Info("call start", zapdriver.OperationStart(operationID.String(), producerID))
 
 		ctx := r.Context()
-		ctx = ctxzap.ToContext(ctx, m.logger)
+		ctx = ctxzap.ToContext(ctx, m.logger.With(
+			zapdriver.OperationCont(operationID.String(), producerID),
+		))
 
 		defer func() {
 			if rcvr := recover(); rcvr != nil {
@@ -38,8 +43,8 @@ func (m *HTTPMiddleware) Handle(next http.Handler) http.Handler {
 		next.ServeHTTP(sw, r.WithContext(ctx))
 
 		latency := time.Since(start)
-
-		fields := []zapcore.Field{
+		zapcoreLevel := httperror.CodeToLevel(sw.status)
+		ctxzap.Extract(ctx).Check(zapcoreLevel, "call end").Write(
 			zap.Int("status", sw.status),
 			zap.Int("content-length", sw.length),
 			zap.Duration("took", latency),
@@ -47,11 +52,7 @@ func (m *HTTPMiddleware) Handle(next http.Handler) http.Handler {
 			zap.String("remote", r.RemoteAddr),
 			zap.String("request", r.RequestURI),
 			zap.String("method", r.Method),
-		}
-		if requestID != "" {
-			fields = append(fields, zap.String("request-id", requestID))
-		}
-		log.Infof(ctx, "request completed", fields...)
+			zapdriver.OperationEnd(operationID.String(), producerID))
 	})
 }
